requirements_simple.txt
pybind11>=2.6.0
numpy>=1.19.0
Pillow>=8.0.0
fastapi>=0.68.0
uvicorn>=0.15.0
websockets>=10.0
av>=10.0.0